Move rect to core
