+++ src/core/rect.hpp
